+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.3.0] - 2021-03-28
+### Changed
+- Modify webpack.config.js for multiple locales
+- Add locale templates
+
 ## [1.2.0] - 2021-03-28
 ### Changed
 - Update page styles